+++ CHANGELOG.md
@@ -74,3 +74,4 @@
 - Bug #70: Update documentation and improve clarity in `HasData` and `DataProvider` classes (@terabytesoftw)
 - Bug #71: Add test cases for closures returning boolean values (`false`, `true`) in `AttributesProvider` class (@terabytesoftw)
 - Bug #72: Update `HasClass` trait docblock to reference `HasAttributes` for attribute management and update related tests (@terabytesoftw)
+- Enh #73: Add `HasStyle` trait with `style()` method and tests (@terabytesoftw)